Fix regression in writing authorized principals
Commit ac6d38e4b7a0a565077229984a3720b2b6a0688b introduced a regression
in writing of the authorized_principals file, resulting in an empty file.

The function `regeneratePrincipalsKeys` in
`services/asymkey/ssh_key_authorized_principals.go` calls the function
`WriteAuthorizedStringForValidKey` for a PublicKey of type
KeyTypePrincipal, and ssh.ParseAuthorizedKey would always fail.

This commit adds additional logic to this function to restore the
previous behaviour when writing the principals file.

Fixes: 36212

Signed-off-by: Peter Verraedt <peter.verraedt@kuleuven.be>